feat: NavigationBar 디자인 수정 및 Header 구현
