+++ src/App.tsx
@@ -1,11 +1,15 @@
 import './App.css';
+import Header from './components/Header';
 import GlobalLayout from './components/layouts/GlobalLayout';
 import Router from './routes';
 
 const App = () => (
-  <GlobalLayout>
-    <Router />
-  </GlobalLayout>
+  <>
+    <Header />
+    <GlobalLayout>
+      <Router />
+    </GlobalLayout>
+  </>
 );
 
 export default App;
